requirements.txt
python-telegram-bot>=20.0
ebooklib>=0.17
edge-tts>=6.0
aiofiles>=23.0.0
flask>=3.0.0
flask-cors>=4.0.0
qrcode>=7.4.0
pillow>=10.0.0
gunicorn>=21.0.0
azure-cognitiveservices-speech>=1.34.0
boto3>=1.34.0
python-dotenv>=1.0.0
TTS>=0.21.0